64-bit (just to make it build).
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@151 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Frontends/VirtualBox/src/VBoxFrameBuffer.cpp
@@ -62,7 +62,7 @@ STDMETHODIMP VBoxFrameBuffer::COMGETTER(Address) (ULONG *aAddress)
 {
     if (!aAddress)
         return E_POINTER;
-    *aAddress = (ULONG) address();
+    *aAddress = (uintptr_t) address();
     return S_OK;
 }
